Added UI, working functionalities and DB.
The project uses React.js and it is structured in components. I added the UI, using the Prime components library and I used it both for design and to create functionalities regarding the tables of content. The DB is working both for users ('Login' component) and for feedbacks ('Feedbacks' components). There is now also routing in order to navigate through the pages; there is available the functionality that separates the users from the non-users, which have less options in the app (editing/adding feedbacks and so on). The data is saved in the SQLite DB by using the interface's inputs. TODO: add checking mechanism for users in Login page; add new component for user's details with functionalities.